Implemented rough demo UI draft.
The demo UI is heavily work in progress.
It tries to map the user journey from the hook all the way to the final
call to action (not implemented yet).
All colors and shapes are preliminary, the first draft focusses on
code organization and extensibility.
The code can be easily extended to cater to different themes.
New input elements can very easily be added.
New pages/or views in existing pages can easily be added as there is no
global state to keep track of.